requirements.txt
markupsafe~=2.1.3
setuptools~=65.5.1
wheel~=0.38.4
jinja2~=3.1.2
dnspython~=2.4.2
email-validator~=2.1.0
idna~=3.4
